created new tables for weightlifting data model

+++ migrations/20250705_create_test_table.sql
@@ -1,10 +1,8 @@
 CREATE TABLE test_migration (
-    id               INT          GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    id               SMALLINT     GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     name             TEXT         NOT NULL,
     created_date     TIMESTAMP    DEFAULT CURRENT_TIMESTAMP
 );
 
 
 INSERT INTO test_migration (name) VALUES ('danny');
-
-
